Added PKGBUILD and binary generation, refactored file names accordingly
