+++ .gitignore
@@ -5,4 +5,14 @@ __pycache__/
 *.swp
 *.zip
 requirements.txt
-src/version.py
+version.py
+pkging/aur/*/PKGBUILD
+keycov.yml
+keycov-*.tar
+keycov-*.tar.xz
+keycov
+!keycov/
+*.[0-9]
+*.[0-9].*
+keycov-bin
+keycov-bintemp
