+++ pkging/aur/keycov/PKGBUILD.in
@@ -1,0 +1,34 @@
+# Maintainer: Ed Jones <ej@kcza.net>
+# Contributor: Ed Jones <ej@kcza.net>
+
+pkgname=s_pkgname
+pkgver=s_pkgver
+pkgrel=1
+source=('s_pkgname-s_pkgver.tar.xz')
+arch=('x86_64')
+license=('LGPL3')
+pkgdesc='s_pkgdesc'
+url=https://github.com/TheSignPainter98/keycov
+# changelog=ChangeLog
+depends=(s_depends)
+sha256sums=('s_sha256sum')
+sha512sums=('s_sha512sum')
+
+package()
+{
+	cd "$srcdir/$pkgname-$pkgver"
+	install -dm755 "${pkgdir}/usr/bin"
+	install -dm755 "${pkgdir}/usr/share/licenses/${pkgname}"
+	install -dm755 "${pkgdir}/usr/share/man/man1"
+
+	install -Dm755 "${pkgname}" "${pkgdir}/usr/bin/${pkgname}"
+	install -Dm755 LICENSE "${pkgdir}/usr/share/licenses/${pkgname}/LICENSE"
+	install -Dm755 "${pkgname}.1.gz" "${pkgdir}/usr/share/man/man1"
+}
+
+check()
+{
+	cd "$srcdir/$pkgname-$pkgver"
+	./keycov --version >/dev/null
+}
+
