+++ src/keycov.py
@@ -1,11 +1,14 @@
-from .args import args, Namespace, parse_args
-from .text_output import output_as_text
-from .analysis_runner import run_analyses
-from .parse_kle import parse_kle
-from .path import get_json_and_yaml_files
-from .util import dict_union, key_pretty_name, serialise_key
+#!/usr/bin/env python3
+
 from json import dumps as jdump
+from keycov.analysis_runner import run_analyses
+from keycov.args import args, Namespace, parse_args
+from keycov.parse_kle import parse_kle
+from keycov.path import get_json_and_yaml_files
+from keycov.text_output import output_as_text
+from keycov.util import dict_union, key_pretty_name, serialise_key
 from os import linesep
+from sys import argv, exit
 from typing import List, Tuple, Union
 from yaml import dump as ydump
 
@@ -64,3 +67,6 @@ def sanitise_layouts(target_layouts:[dict], input_layouts:[dict]):
         for key in layout[1]:
             key['serialised'] = serialise_key(key)
             key['pretty-name'] = key_pretty_name(key)
+
+if __name__ == '__main__':
+    exit(main(argv))
