+++ src/keycov/version.py.in
@@ -1,6 +1,7 @@
 version_notice: str = '''
-KeyCov VERSION, an automatic keycap kitting analyser
+s_name s_version
 Copyright (C) 2020 Edward Jones
+s_desc.
 
 This program is free software: you can redistribute it and/or modify it under
 the terms of the GNU Lesser General Public License as published by the Free
